+++ CHANGELOG.md
@@ -19,7 +19,8 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 
 ### Fixed
 
-- Subsequent calls to `next_back` in `CigletIterator` now properly return `None` after invalid state is reached 
+- Subsequent calls to `next_back` in `CigletIterator` now properly return `None` after invalid state is reached
+- Methods in `StatesSequence` and `StatesSequenceMut` now properly skip empty `Ciglet`s (with an increment of 0)
 
 ## [0.0.27] - 2026-04-15